Move the tests to test_pandas.py and correct the skip for Pandas 0.23
