+++ python/pyarrow/tests/test_array.py
@@ -36,7 +36,6 @@ import pytz
 
 import pyarrow as pa
 import pyarrow.tests.strategies as past
-from pyarrow.vendored.version import Version
 
 
 def test_total_bytes_allocated():
@@ -263,69 +262,6 @@ def test_to_pandas_zero_copy():
         base_refcount = sys.getrefcount(series.values.base)
         assert base_refcount == 2
         series.sum()
-
-
-@pytest.mark.skipif(
-    Version('0.23') <= Version(pa.__version__) < Version('1.0.0'),
-    reason='Extension types introduced in later versions of Pandas')
-@pytest.mark.pandas
-def test_to_pandas_types_mapper():
-    # https://issues.apache.org/jira/browse/ARROW-9664
-    import pandas as pd
-    data = pa.array([1, 2, 3], pa.int64())
-
-    # Test with mapper function
-    types_mapper = {pa.int64(): pd.Int64Dtype()}.get
-    result = data.to_pandas(types_mapper=types_mapper)
-    assert result.dtype == types_mapper(data.type)
-
-    # Test mapper function returning None
-    types_mapper = {pa.int64(): None}.get
-    result = data.to_pandas(types_mapper=types_mapper)
-    assert result.dtype == data.type.to_pandas_dtype()
-
-    # Test mapper function not containing the dtype
-    types_mapper = {pa.float64(): pd.Float64Dtype()}.get
-    result = data.to_pandas(types_mapper=types_mapper)
-    assert result.dtype == data.type.to_pandas_dtype()
-
-    # Test for the interval extension dtype
-    # -> ignores mapping and uses default conversion
-    types_mapper = {pa.float64(): pd.IntervalDtype()}.get
-    result = data.to_pandas(types_mapper=types_mapper)
-    assert result.dtype == data.type.to_pandas_dtype()
-
-
-@pytest.mark.skipif(
-    Version('0.23') <= Version(pa.__version__) < Version('1.0.0'),
-    reason='Extension types introduced in later versions of Pandas')
-@pytest.mark.pandas
-def test_chunked_array_to_pandas_types_mapper():
-    # https://issues.apache.org/jira/browse/ARROW-9664
-    import pandas as pd
-    data = pa.chunked_array([pa.array([1, 2, 3], pa.int64())])
-    assert isinstance(data, pa.ChunkedArray)
-
-    # Test with mapper function
-    types_mapper = {pa.int64(): pd.Int64Dtype()}.get
-    result = data.to_pandas(types_mapper=types_mapper)
-    assert result.dtype == types_mapper(data.type)
-
-    # Test mapper function returning None
-    types_mapper = {pa.int64(): None}.get
-    result = data.to_pandas(types_mapper=types_mapper)
-    assert result.dtype == data.type.to_pandas_dtype()
-
-    # Test mapper function not containing the dtype
-    types_mapper = {pa.float64(): pd.Float64Dtype()}.get
-    result = data.to_pandas(types_mapper=types_mapper)
-    assert result.dtype == data.type.to_pandas_dtype()
-
-    # Test for the interval extension dtype
-    # -> ignores mapping and uses default conversion
-    types_mapper = {pa.float64(): pd.IntervalDtype()}.get
-    result = data.to_pandas(types_mapper=types_mapper)
-    assert result.dtype == data.type.to_pandas_dtype()
 
 
 @pytest.mark.nopandas
